+++ src/clj/qbits/spandex/url.clj
@@ -1,0 +1,58 @@
+(ns qbits.spandex.url)
+
+(declare x-fragment-interpose-comma
+         x-fragment-interpose-slash)
+
+(defprotocol URL
+  (encode [x]))
+
+(defprotocol URLFragment
+  (encode-fragment [value]))
+
+(defn ^:no-doc string-builder
+  ([] (StringBuilder.))
+  ([^StringBuilder sb x] (.append sb x))
+  ([^StringBuilder sb] (.toString sb)))
+
+(defn ^:no-doc url-string-builder
+  ([] (StringBuilder. "/"))
+  ([^StringBuilder sb x] (.append sb x))
+  ([^StringBuilder sb] (.toString sb)))
+
+(extend-protocol URLFragment
+  clojure.lang.Sequential
+  (encode-fragment [value]
+    ;; multi index fragment
+    (transduce x-fragment-interpose-comma string-builder value))
+
+  clojure.lang.Keyword
+  (encode-fragment [value] (name value))
+
+  String
+  (encode-fragment [s] s)
+
+  Object
+  (encode-fragment [value] (str value)))
+
+(def ^:no-doc x-fragment-interpose-comma
+  (comp (remove nil?)
+        (map encode-fragment)
+        (interpose ",")))
+
+(def ^:no-doc x-fragment-interpose-slash
+  (comp (remove nil?)
+        (map encode-fragment)
+        (interpose "/")))
+
+(extend-protocol URL
+  String
+  (encode [s] s)
+
+  Object
+  (encode [parts]
+    (transduce x-fragment-interpose-slash
+               url-string-builder
+               parts))
+
+  nil
+  (encode [_] "/"))